+++ src/util/printer.rs
@@ -194,6 +194,7 @@ impl Printable for nodes::Expression<'_> {
             Self::Binary(node) => node.print_ast(indent),
             Self::FunctionCall(node) => node.print_ast(indent),
             Self::Sizeof(node) => node.print_ast(indent),
+            Self::As(_, _) => todo!("Expression::As.print_ast()"),
         }
     }
 }